NXCM-4928: switch to junit-411
(with latest litmus)

Most interesting part here was the fact that JUnit 4.11
deliberately changes test execution order, and it turned out
that ProxyServerTest isolation actually depended on old
ordering (or new ordering broke it, does not matter).

Reason was the UrlConnection and the fact it caches the
authc info. Hence, with new test ordering, 200 response
was received (as same authc info was already cached) instead
of the expected 401.

For now, an "illegal" workaround added, with Sun
private classes to simply clear caches between test
executions.